Add Parser to handle commands
