+++ src/main/java/Storage.java
@@ -3,7 +3,6 @@ import jdk.jshell.spi.ExecutionControl;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
